requirements.txt
asgiref==3.8.1
charset-normalizer==3.4.2
crispy-bootstrap5==2025.6
Django==5.2.3
django-crispy-forms==2.4
django-widget-tweaks==1.5.0
pillow==11.2.1
python-decouple==3.8
reportlab==4.4.2
sqlparse==0.5.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
